+++ tests/test_pending_experiments.py
@@ -7,7 +7,6 @@ import pandas as pd
 import pytest
 from pytest import param
 
-from baybe._optional.info import SCIKIT_LEARN_EXTRA_INSTALLED
 from baybe.acquisition.base import AcquisitionFunction
 from baybe.exceptions import IncompatibleAcquisitionFunctionError, UnusedObjectWarning
 from baybe.recommenders import (
@@ -41,10 +40,6 @@ _hybrid_params = ["Categorical_1", "Num_disc_1", "Conti_finite1", "Conti_finite2
             _discrete_params,
             PAMClusteringRecommender(),
             id="pam_discrete",
-            marks=pytest.mark.skipif(
-                not SCIKIT_LEARN_EXTRA_INSTALLED,
-                reason="Optional dependency 'scikit-learn-extra' not installed",
-            ),
         ),
         param(
             _discrete_params,